requirements.txt
pandas
pdfminer.six
nltk
scikit-learn
sqlite3-binary